% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youngSwimmers-package.R
\docType{package}
\name{youngSwimmers-package}
\alias{youngSwimmers}
\alias{youngSwimmers-package}
\title{youngSwimmers}
\description{
Data from the young elite swimmers study.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NIM-ACh/youngSwimmers/}
  \item \url{https://nim-ach.github.io/youngSwimmers/}
  \item Report bugs at \url{https://github.com/NIM-ACh/youngSwimmers/issues/}
}

}
\author{
\strong{Maintainer}: Matías Castillo Aguilar \email{matcasti@umag.cl} (\href{https://orcid.org/0000-0001-7291-247X}{ORCID})

Authors:
\itemize{
  \item Pablo Valdés Badilla
  \item Tomás Herrera Valenzuela
  \item Eduardo Guzmán Muñoz
  \item Pedro Delgado Floody
  \item David Cristóbal Andrade
  \item Michele M. Moraes
  \item Rosa M. E. Arantes
  \item Cristian Núñez Espinosa
}

}
\keyword{internal}
